+++ mojo/edk/system/ports/port.cc
@@ -8,8 +8,8 @@ namespace mojo {
 namespace edk {
 namespace ports {
 
-Port::Port(uint32_t next_sequence_num_to_send,
-           uint32_t next_sequence_num_to_receive)
+Port::Port(uint64_t next_sequence_num_to_send,
+           uint64_t next_sequence_num_to_receive)
     : state(kUninitialized),
       next_sequence_num_to_send(next_sequence_num_to_send),
       last_sequence_num_to_receive(0),
